+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 > Vue Next 文档站时间线版见 [更新日志](https://ant-design-x-vue-next-524.netlify.app/development/changelog.html)。
 
+## 1.21.0
+
+`2026-07-30`
+
+### ant-design-x-vue-next
+
+- 🆕 Conversations.Creation / `creation` 新建对话
+- 📖 Conversations 文档更新
+- 🛠 发布 npm `ant-design-x-vue-next@1.21.0`
+
 ## 1.20.0
 
 `2026-07-30`